+++ shared/src/main/scala/ky/korins/blake3/RFC4648.scala
@@ -62,17 +62,17 @@ object RFC4648 {
       case 2 =>
         sb append alphabet((bytes(offset) >>> 2) & 0x3f)
         sb append alphabet(
-          ((bytes(offset) & 0x3) << 4) | ((bytes(1) >>> 4) & 0xf)
+          ((bytes(offset) & 0x3) << 4) | ((bytes(offset + 1) >>> 4) & 0xf)
         )
         sb append alphabet((bytes(offset + 1) & 0xf) << 2)
 
       case _ => // 3 or more
         sb append alphabet((bytes(offset) >>> 2) & 0x3f)
         sb append alphabet(
-          ((bytes(offset) & 0x3) << 4) | ((bytes(1) >>> 4) & 0xf)
+          ((bytes(offset) & 0x3) << 4) | ((bytes(offset + 1) >>> 4) & 0xf)
         )
         sb append alphabet(
-          ((bytes(offset + 1) & 0xf) << 2) | ((bytes(2) >>> 6) & 0x3)
+          ((bytes(offset + 1) & 0xf) << 2) | ((bytes(offset + 2) >>> 6) & 0x3)
         )
         sb append alphabet(bytes(offset + 2) & 0x3f)
     }
@@ -95,7 +95,7 @@ object RFC4648 {
       case 2 =>
         sb append alphabet((bytes(offset) >>> 3) & 0x1f)
         sb append alphabet(
-          ((bytes(offset) & 0x7) << 2) | ((bytes(1) >>> 6) & 0x3)
+          ((bytes(offset) & 0x7) << 2) | ((bytes(offset + 1) >>> 6) & 0x3)
         )
         sb append alphabet((bytes(offset + 1) >>> 1) & 0x1f)
         sb append alphabet((bytes(offset + 1) & 0x1) << 4)
@@ -103,25 +103,25 @@ object RFC4648 {
       case 3 =>
         sb append alphabet((bytes(offset) >>> 3) & 0x1f)
         sb append alphabet(
-          ((bytes(offset) & 0x7) << 2) | ((bytes(1) >>> 6) & 0x3)
+          ((bytes(offset) & 0x7) << 2) | ((bytes(offset + 1) >>> 6) & 0x3)
         )
         sb append alphabet((bytes(offset + 1) >>> 1) & 0x1f)
         sb append alphabet(
-          ((bytes(offset + 1) & 0x1) << 4) | ((bytes(2) >>> 4) & 0xf)
+          ((bytes(offset + 1) & 0x1) << 4) | ((bytes(offset + 2) >>> 4) & 0xf)
         )
         sb append alphabet((bytes(offset + 2) & 0xf) << 1)
 
       case 4 =>
         sb append alphabet((bytes(offset) >>> 3) & 0x1f)
         sb append alphabet(
-          ((bytes(offset) & 0x7) << 2) | ((bytes(1) >>> 6) & 0x3)
+          ((bytes(offset) & 0x7) << 2) | ((bytes(offset + 1) >>> 6) & 0x3)
         )
         sb append alphabet((bytes(offset + 1) >>> 1) & 0x1f)
         sb append alphabet(
-          ((bytes(offset + 1) & 0x1) << 4) | ((bytes(2) >>> 4) & 0xf)
+          ((bytes(offset + 1) & 0x1) << 4) | ((bytes(offset + 2) >>> 4) & 0xf)
         )
         sb append alphabet(
-          ((bytes(offset + 2) & 0xf) << 1) | ((bytes(3) >>> 7) & 0x1)
+          ((bytes(offset + 2) & 0xf) << 1) | ((bytes(offset + 3) >>> 7) & 0x1)
         )
         sb append alphabet((bytes(offset + 3) >>> 2) & 0x1f)
         sb append alphabet((bytes(offset + 3) & 0x3) << 3)
@@ -129,18 +129,18 @@ object RFC4648 {
       case _ => // 5 or more
         sb append alphabet((bytes(offset) >>> 3) & 0x1f)
         sb append alphabet(
-          ((bytes(offset) & 0x7) << 2) | ((bytes(1) >>> 6) & 0x3)
+          ((bytes(offset) & 0x7) << 2) | ((bytes(offset + 1) >>> 6) & 0x3)
         )
         sb append alphabet((bytes(offset + 1) >>> 1) & 0x1f)
         sb append alphabet(
-          ((bytes(offset + 1) & 0x1) << 4) | ((bytes(2) >>> 4) & 0xf)
+          ((bytes(offset + 1) & 0x1) << 4) | ((bytes(offset + 2) >>> 4) & 0xf)
         )
         sb append alphabet(
-          ((bytes(offset + 2) & 0xf) << 1) | ((bytes(3) >>> 7) & 0x1)
+          ((bytes(offset + 2) & 0xf) << 1) | ((bytes(offset + 3) >>> 7) & 0x1)
         )
         sb append alphabet((bytes(offset + 3) >>> 2) & 0x1f)
         sb append alphabet(
-          ((bytes(offset + 3) & 0x3) << 3) | ((bytes(4) >>> 5) & 0x7)
+          ((bytes(offset + 3) & 0x3) << 3) | ((bytes(offset + 4) >>> 5) & 0x7)
         )
         sb append alphabet(bytes(offset + 4) & 0x1f)
     }